requirements.txt
azure-eventhub==1.3.2
pandas==0.25.1